Moved functionality for choosing HERA observing season to a new script, hera_season.py. Modified noise.py and foregrounds.py accordingly. Performed light testing, and modules seem to work as expected.
