+++ hera_sim/foregrounds.py
@@ -11,7 +11,7 @@ from builtins import range
 
 from . import noise
 from . import utils
-
+from .hera_season import get_season
 
 def diffuse_foreground(lsts, fqs, bl_vec, Tsky_mdl=None, omega_p=None, season=None,
                        standoff=0.0, delay_filter_type='tophat', delay_filter_normalize=None,
@@ -45,7 +45,7 @@ def diffuse_foreground(lsts, fqs, bl_vec, Tsky_mdl=None, omega_p=None, season=No
     if Tsky_mdl is None:
         raise TypeError("Tsky_mdl is a required parameter of diffuse_foreground")
     if omega_p is None:
-        seas = noise.get_season(season)
+        seas = get_season(season)
         omega_p = seas.noise.get_omega_p(fqs)
 
     # generate a Tsky visibility in time and freq space, convert from K to Jy
